feat: upgrade to react-router-dom v6, other component upgrades
