+++ ui/package.json
@@ -53,7 +53,7 @@
     "raw-loader": "^4.0.2",
     "react-axe": "^3.5.4",
     "react-docgen-typescript-loader": "^3.7.2",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "^6.0.0",
     "regenerator-runtime": "^0.14.0",
     "rimraf": "^6.0.0",
     "style-loader": "^4.0.0",
@@ -72,7 +72,7 @@
     "webpack-merge": "^6.0.0"
   },
   "dependencies": {
-    "@melloware/react-logviewer": "^5.1.1",
+    "@melloware/react-logviewer": "^6.0.0",
     "@patternfly/react-core": "6.0.0",
     "@patternfly/react-icons": "6.0.0",
     "@patternfly/react-log-viewer": "6.0.0",
