+++ ui/src/app/components/Pages/AppLayout/AppLayout.tsx
@@ -43,7 +43,7 @@ const AppLayout: React.FunctionComponent<IAppLayout> = ({ children }) => {
 
   const renderNavItem = (route: IAppRoute, index: number) => (
     <NavItem key={`${route.label}-${index}`} id={`${route.label}-${index}`} isActive={route.path === location.pathname}>
-      <NavLink exact={route.exact} to={route.path}>
+      <NavLink to={route.path}>
         {route.label}
       </NavLink>
     </NavItem>
